Meeting Summarizer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
streamlit==1.38.0
requests==2.32.3
plotly==5.24.1
pandas==2.2.2
numpy==1.26.4
librosa==0.10.2.post1
noisereduce==3.0.2
scipy==1.11.4
soundfile==0.12.1
icalendar==5.0.13
websockets==12.0

